fix controller
