+++ src/main/java/com/example/OsmsApplication.java
@@ -1,0 +1,13 @@
+package com.example;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OsmsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OsmsApplication.class, args);
+	}
+
+}
