+++ src/main/java/com/example/services/GreetingService.java
@@ -5,5 +5,9 @@ import org.springframework.stereotype.Service;
 @Service
 public class GreetingService
 {
+  
+  public String helloWorld() {
+    return "Hello World from OpenShift";
+  }
 
 }
